+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [Library] Icon flip option for `text input` component trailing action (Orange-OpenSource/ouds-ios#1041)
 - [Library] Icon flip option for `button` component (Orange-OpenSource/ouds-ios#1124)
 - [DesignToolbox] Support of `watchOS`
 - [DesignToolbox] Support of `tvOS`